# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "專案名稱 API",
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Home page",
                "operationId": "ac5f32eec336b092a10a014ae5e0d4ed",
                "responses": {
                    "default": {
                        "description": "Welcome page"
                    }
                }
            }
        },
        "/api/task/{taskId}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete Task",
                "description": "刪除task",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "需要任務id來刪除",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "ajax 帶 api token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid taskId supplied"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "Authorization": {
                "name": "Authorization",
                "in": "header",
                "description": "ajax请求要加上X-Requested-With:XMLHttpRequest头字段",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    },
    "security": [
        []
    ]
}